Renames file to `index`
